% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row.R
\name{add_row}
\alias{add_row}
\title{add_row}
\usage{
add_row(row, dbcon, tbl_name, tbl_fields, user_id = NULL,
  commit = FALSE)
}
\arguments{
\item{row}{single row from google sheet (tibble of dim 1 x n)}

\item{dbcon}{Database connection, as created by [RPostgres::dbConnect]}

\item{tbl_name}{name of BETYdb table (character, length 1)}

\item{tbl_fields}{names of fields of BETYdb table that are also field names of row;
used to subset row (character vector)}

\item{user_id}{BETYdb user id; only needed when inserting new experiments;
default = NULL (integer length 1)}

\item{commit}{value indicating whether to commit database transaction;
(boolean: TRUE or FALSE, default is FALSE)}
}
\value{
exits on success
}
\description{
Add a row of data from google sheet into specified BETYdb table
}
\examples{
\dontrun{
add_row(row = tibble(name = 'MAC Season 8: Border', start_date = '2019-01-01',
end_date = '2019-03-31'), dbcon = dbcon, tbl_name = 'experiments',
tbl_fields = c('name', 'start_date', 'end_date', 'design', 'description'), user_id = 6000000004,
commit = FALSE)
}
}
